feat: Add Docker Compose support, improve security, and apply minor fixes

+++ library/src/main/java/org/pollub/library/user/service/UserService.java
@@ -7,6 +7,7 @@ import org.pollub.library.user.model.Role;
 import org.pollub.library.user.model.RoleSetDto;
 import org.pollub.library.user.model.User;
 import org.pollub.library.user.repository.IUserRepository;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -51,10 +52,18 @@ public class UserService implements IUserService {
 
     @Override
     public void deleteUserById(Long id) {
+        if (id.equals(getCurrentUserId())) {
+            throw new IllegalArgumentException("You cannot delete yourself");
+        }
+
         if (!userRepository.existsById(id)) {
             throw new UserNotFoundException("User with ID " + id + " not found");
         }
         userRepository.deleteById(id);
+    }
+
+    private Long getCurrentUserId() {
+        return ((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
     }
 
 }
